bots/telegram/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nmtelegrambot"
version = "2.2.0"
description = "Telegram runtime for NeverMine"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot[webhooks]>=21,<22",
  "python-dotenv>=1,<2",
  "httpx>=0.27,<1",
  "redis>=5,<6",
  "psycopg[binary]>=3,<4"
]

[tool.setuptools.packages.find]
include = ["nmbot*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["N818"]

[project.optional-dependencies]
qa = [
  "pytest>=8,<9",
  "redis>=5,<6",
  "psycopg[binary]>=3,<4"
]
